Revert src/lzma.js
