+++ package.json
@@ -22,7 +22,8 @@
     "karma-mocha": "^1.3.0",
     "mocha": "^3.2.0",
     "rollup": "^1.31.0",
-    "rollup-plugin-terser": "^5.2.0"
+    "rollup-plugin-terser": "^5.2.0",
+    "terser": "^4.6.3"
   },
   "scripts": {
     "build": "node rollup.js",
