+++ src/lzma-d.mjs
@@ -1,21 +1,15 @@
-import {
-    action_decompress,
-    decompress
-} from './lzma_worker';
+import { decompress, action_decompress } from "./lzma-algo";
 
-export {
-    decompress
-};
+export { decompress };
 
 export function LZMA() {}
-
 LZMA["decompress"] = decompress;
 LZMA.prototype["decompress"] = decompress;
 
 if (typeof self != "undefined" && 'importScripts' in self) {
     addEventListener("message", function (e) {
         if (e["data"]["action"] == action_decompress) {
-            LZMA.decompress(e["data"]["data"], e["data"]["cbn"]);
+            decompress(e["data"]["data"], e["data"]["cbn"]);
         }
-    })
+    });
 }
