+++ src/lzma-min.js
@@ -1,1 +1,1 @@
-!function(r,n){"object"==typeof exports&&"undefined"!=typeof module?n(exports):"function"==typeof define&&define.amd?define(["exports"],n):n((r=r||self).LZMA={})}(this,(function(r){"use strict";var n="function"==typeof setImmediate?setImmediate:setTimeout,f=[4294967295,-4294967296],t=[0,-0x8000000000000000],o=[0,0],u=[1,0];function i(r,n){postMessage({action:3,cbn:n,result:r})}function e(r){var n=[];return n[r-1]=void 0,n}function c(r,n){return l(r[0]+n[0],r[1]+n[1])}function a(r,n){return function(r,n){var f;f=n,0>n&&(f+=4294967296);return[f,4294967296*r]}(~~Math.max(Math.min(r[1]/4294967296,2147483647),-2147483648)&~~Math.max(Math.min(n[1]/4294967296,2147483647),-2147483648),h(r)&h(n))}function v(r,n){var f,t;return r[0]==n[0]&&r[1]==n[1]?0:(t=0>n[1],(f=0>r[1])&&!t?-1:!f&&t?1:0>m(r,n)[1]?-1:1)}function l(r,n){var f,t;for(n=(n%=0x10000000000000000)-(f=n%4294967296)+(t=4294967296*Math.floor((r%=0x10000000000000000)/4294967296)),r=r-t+f;0>r;)r+=4294967296,n-=4294967296;for(;r>4294967295;)r-=4294967296,n+=4294967296;for(n%=0x10000000000000000;n>0x7fffffff00000000;)n-=0x10000000000000000;for(;-0x8000000000000000>n;)n+=0x10000000000000000;return[r,n]}function s(r,n){return r[0]==n[0]&&r[1]==n[1]}function d(r){return 0>r?[r+4294967296,-4294967296]:[r,0]}function h(r){return 2147483648>r[0]?~~Math.max(Math.min(r[0],2147483647),-2147483648):~~Math.max(Math.min(r[0]-4294967296,2147483647),-2147483648)}function p(r){return r>30?p(30)*p(r-30):1<<r}function w(r,n){var f,u,i,e;if(n&=63,s(r,t))return n?o:r;if(0>r[1])throw Error("Neg");return e=p(n),u=r[1]*e%0x10000000000000000,0x8000000000000000>(u+=f=(i=r[0]*e)-i%4294967296)||(u-=0x10000000000000000),[i-=f,u]}function M(r,n){var f;return f=p(n&=63),l(Math.floor(r[0]/f),r[1]/f)}function m(r,n){return l(r[0]-n[0],r[1]-n[1])}function b(r,n){return r.t=n,r.o=0,r.count=n.length,r}function x(r){return r.count>r.o?255&r.t[r.o++]:-1}function g(r,n,f,t){return r.count>r.o?(D(r.t,r.o,n,f,t=Math.min(t,r.count-r.o)),r.o+=t,t):-1}function y(r){return r.t=e(32),r.count=0,r}function E(r){var n=r.t;return n.length=r.count,n}function k(r,n){r.t[r.count++]=n<<24>>24}function S(r,n,f,t){D(n,f,r.t,r.count,t),r.count+=t}function D(r,n,f,t,o){for(var u=0;o>u;++u)f[t+u]=r[n+u]}function j(r,n,t,u,i){var c,a;if(0>v(u,f))throw Error("invalid length "+u);for(r.u=u,function(r,n){!function(r,n){r.i=n;for(var f=0;n>1<<f;++f);r.v=2*f}(n,1<<r.l),n.s=r.f,function(r,n){var f=r.h;r.h=n,r.p&&f!=r.h&&(r.M=-1,r.p=null)}(n,r.m),n.g=0,n.k=3,n.S=2,n.D=3}(i,c=function(r){var n;for(r.j=e(4),r.I=[],r.N={},r.$=e(192),r._=e(12),r.L=e(12),r.O=e(12),r.T=e(12),r.q=e(192),r.A=[],r.B=e(114),r.C=Er({},4),r.F=sr({}),r.G=sr({}),r.H={},r.J=[],r.K=[],r.P=[],r.R=e(16),r.U=e(4),r.V=e(4),r.W=[o],r.X=[o],r.Y=[0],r.properties=e(5),r.Z=e(128),r.rr=0,r.h=1,r.nr=0,r.tr=-1,r.or=0,n=0;4096>n;++n)r.I[n]={};for(n=0;4>n;++n)r.A[n]=Er({},6);return r}({})),function(r,n){r.properties[0]=9*(5*r.S+r.g)+r.k<<24>>24;for(var f=0;4>f;++f)r.properties[1+f]=r.i>>8*f<<24>>24;S(n,r.properties,0,5)}(c,t),a=0;64>a;a+=8)k(t,255&h(M(u,a)));r.ur=(c.ir=0,c.er=n,c.cr=0,function(r){var n,f;r.p||(f=4,r.h||(f=2),function(r,n){r.ar=n>2,r.ar?(r.vr=0,r.lr=4,r.sr=66560):(r.vr=2,r.lr=3,r.sr=0)}(n={},f),r.p=n);if(function(r,n,f){var t,o;if(null!=r.dr&&r.hr==f&&r.pr==n)return;for(r.pr=n,r.wr=(1<<n)-1,r.hr=f,r.dr=e(o=1<<r.hr+r.pr),t=0;o>t;++t)r.dr[t]=mr({})}(r.H,r.g,r.k),r.i==r.M&&r.tr==r.s)return;(function(r,n,f,t,o){var u,i;1073741567>n&&(r.Mr=16+(t>>1),function(r,n,f,t){var o;r.mr=n,r.br=f,o=n+f+t,(null==r.xr||r.gr!=o)&&(r.xr=null,r.gr=o,r.xr=e(r.gr));r.yr=r.gr-f}(r,n+f,t+o,256+~~((n+f+t+o)/2)),r.Er=t,r.kr!=(u=n+1)&&(r.Sr=e(2*(r.kr=u))),i=65536,r.ar&&(i=n-1,i|=i>>1,i|=i>>2,i|=i>>4,i|=i>>8,i>>=1,(i|=65535)>16777216&&(i>>=1),r.Dr=i,i+=1,i+=r.sr),i!=r.jr&&(r.Ir=e(r.jr=i)))})(r.p,r.i,4096,r.s,274),r.M=r.i,r.tr=r.s}(c),c.N.Nr=t,function(r){(function(r){r.$r=0,r._r=0;for(var n=0;4>n;++n)r.j[n]=0})(r),function(r){r.Lr=o,r.Or=o,r.Range=-1,r.Tr=1,r.qr=0}(r.N),_r(r.$),_r(r.q),_r(r._),_r(r.L),_r(r.O),_r(r.T),_r(r.B),function(r){var n,f=1<<r.hr+r.pr;for(n=0;f>n;++n)_r(r.dr[n].zr)}(r.H);for(var n=0;4>n;++n)_r(r.A[n].Ar);ar(r.F,1<<r.S),ar(r.G,1<<r.S),_r(r.C.Ar),r.Br=0,r.Cr=0,r.Fr=0,r.Gr=0}(c),rr(c),Z(c),c.F.Hr=c.s+1-2,hr(c.F,1<<c.S),c.G.Hr=c.s+1-2,hr(c.G,1<<c.S),c.Jr=o,function(r,n){return r.Kr=n,r.Pr=null,r.Qr=1,r}({},c))}function I(r,n,f){return r.Rr=y({}),j(r,b({},n),r.Rr,d(n.length),f),r}function N(r,n,t){var u,i,c,a,v="",l=[];for(i=0;5>i;++i){if(-1==(c=x(n)))throw Error("truncated input");l[i]=c<<24>>24}if(!function(r,n){var f,t,o,u,i,c,a;if(5>n.length)return 0;for(o=(a=255&n[0])%9,u=(c=~~(a/9))%5,i=~~(c/5),f=0,t=0;4>t;++t)f+=(255&n[1+t])<<8*t;if(f>99999999||!function(r,n,f,t){if(n>8||f>4||t>4)return 0;!function(r,n,f){var t,o;if(null!=r.dr&&r.hr==f&&r.pr==n)return;for(r.pr=n,r.wr=(1<<n)-1,r.hr=f,r.dr=e(o=1<<r.hr+r.pr),t=0;o>t;++t)r.dr[t]=W({})}(r.Ur,f,n);var o=1<<t;return Q(r.Vr,o),Q(r.Wr,o),r.Xr=o-1,1}(r,o,u,i))return 0;return function(r,n){if(0>n)return 0;r.Yr!=n&&(r.Yr=n,r.Zr=Math.max(r.Yr,1),function(r,n){null!=r.rn&&r.nn==n||(r.rn=e(n));r.nn=n,r.fn=0,r.tn=0}(r.on,Math.max(r.Zr,4096)));return 1}(r,f)}(u=function(r){r.on={},r.un={},r.in=e(192),r.en=e(12),r.cn=e(12),r.an=e(12),r.vn=e(12),r.ln=e(192),r.sn=e(4),r.dn=e(114),r.hn=gr({},4),r.Vr=U({}),r.Wr=U({}),r.Ur={};for(var n=0;4>n;++n)r.sn[n]=gr({},6);return r}({}),l))throw Error("corrupted input");for(i=0;64>i;i+=8){if(-1==(c=x(n)))throw Error("truncated input");1==(c=c.toString(16)).length&&(c="0"+c),v=c+""+v}r.u=/^0+$|^f+$/i.test(v)?f:(a=parseInt(v,16))>4294967295?f:d(a),r.ur=function(r,n,f,t){return r.un.Nr=n,G(r.on),r.on.pn=f,function(r){r.on.tn=0,r.on.fn=0,_r(r.in),_r(r.ln),_r(r.en),_r(r.cn),_r(r.an),_r(r.vn),_r(r.dn),function(r){var n,f;for(f=1<<r.hr+r.pr,n=0;f>n;++n)_r(r.dr[n].wn)}(r.Ur);for(var n=0;4>n;++n)_r(r.sn[n].Ar);V(r.Vr),V(r.Wr),_r(r.hn.Ar),function(r){r.Mn=0,r.Range=-1;for(var n=0;5>n;++n)r.Mn=r.Mn<<8|x(r.Nr)}(r.un)}(r),r.state=0,r.mn=0,r.bn=0,r.xn=0,r.gn=0,r.yn=t,r.Jr=o,r.En=0,function(r,n){return r.Pr=n,r.Kr=null,r.Qr=1,r}({},r)}(u,n,t,r.u)}function $(r,n){return r.Rr=y({}),N(r,b({},n),r.Rr),r}function _(r,n){return r.xr[r.kn+r.fn+n]}function L(r,n,f,t){var o,u;for(r.Sn&&r.fn+n+t>r.tn&&(t=r.tn-(r.fn+n)),++f,u=r.kn+r.fn+n,o=0;t>o&&r.xr[u+o]==r.xr[u+o-f];++o);return o}function O(r){return r.tn-r.fn}function T(r){var n,f;if(!r.Sn)for(;;){if(!(f=-r.kn+r.gr-r.tn))return;if(-1==(n=g(r.pn,r.xr,r.kn+r.tn,f)))return r.Dn=r.tn,r.kn+r.Dn>r.yr&&(r.Dn=r.yr-r.kn),void(r.Sn=1);r.tn+=n,r.fn+r.br>r.tn||(r.Dn=r.tn-r.br)}}function q(r,n){r.kn+=n,r.Dn-=n,r.fn-=n,r.tn-=n}var z=function(){var r,n,f,t=[];for(r=0;256>r;++r){for(f=r,n=0;8>n;++n)0!=(1&f)?(f>>>=1,f^=-306674912):f>>>=1;t[r]=f}return t}();function A(r){var n;(r.jn+=1)<r.kr||(r.jn=0),function(r){r.fn+=1,r.fn>r.Dn&&(r.kn+r.fn>r.yr&&function(r){var n,f,t;for((t=r.kn+r.fn-r.mr)>0&&--t,f=r.kn+r.tn-t,n=0;f>n;++n)r.xr[n]=r.xr[t+n];r.kn-=t}(r),T(r))}(r),1073741823==r.fn&&(B(r.Sr,2*r.kr,n=r.fn-r.kr),B(r.Ir,r.jr,n),q(r,n))}function B(r,n,f){var t,o;for(t=0;n>t;++t)(o=r[t]||0)>f?o-=f:o=0,r[t]=o}function C(r){var n=r.fn-r.tn;n&&(S(r.pn,r.rn,r.tn,n),r.nn>r.fn||(r.fn=0),r.tn=r.fn)}function F(r,n){var f=r.fn-n-1;return 0>f&&(f+=r.nn),r.rn[f]}function G(r){C(r),r.pn=null}function H(r){return 4>(r-=2)?r:3}function J(r){return 4>r?0:10>r?r-3:r-6}function K(r){if(!r.Qr)throw Error("bad state");if(r.Kr)throw Error("No encoding");return function(r){var n=function(r){var n,f,t,o,i,e;if(e=h(r.Jr)&r.Xr,$r(r.un,r.in,(r.state<<4)+e)){if($r(r.un,r.en,r.state))t=0,$r(r.un,r.cn,r.state)?($r(r.un,r.an,r.state)?($r(r.un,r.vn,r.state)?(f=r.gn,r.gn=r.xn):f=r.xn,r.xn=r.bn):f=r.bn,r.bn=r.mn,r.mn=f):$r(r.un,r.ln,(r.state<<4)+e)||(r.state=7>r.state?9:11,t=1),t||(t=R(r.Wr,r.un,e)+2,r.state=7>r.state?8:11);else if(r.gn=r.xn,r.xn=r.bn,r.bn=r.mn,t=2+R(r.Vr,r.un,e),r.state=7>r.state?7:10,4>(i=yr(r.sn[H(t)],r.un)))r.mn=i;else if(r.mn=(2|1&i)<<(o=(i>>1)-1),14>i)r.mn+=function(r,n,f,t){var o,u,i=1,e=0;for(u=0;t>u;++u)o=$r(f,r,n+i),i<<=1,i+=o,e|=o<<u;return e}(r.dn,r.mn-i-1,r.un,o);else if(r.mn+=function(r,n){var f,t,o=0;for(f=n;0!=f;f-=1)r.Range>>>=1,r.Mn-=r.Range&(t=r.Mn-r.Range>>>31)-1,o=o<<1|1-t,-16777216&r.Range||(r.Mn=r.Mn<<8|x(r.Nr),r.Range<<=8);return o}(r.un,o-4)<<4,r.mn+=function(r,n){var f,t,o=1,u=0;for(t=0;r.In>t;++t)f=$r(n,r.Ar,o),o<<=1,o+=f,u|=f<<t;return u}(r.hn,r.un),0>r.mn)return-1==r.mn?1:-1;if(v(d(r.mn),r.Jr)>=0||r.mn>=r.Zr)return-1;!function(r,n,f){var t=r.fn-n-1;for(0>t&&(t+=r.nn);0!=f;f-=1)r.nn>t||(t=0),r.rn[r.fn]=r.rn[t],r.fn+=1,t+=1,r.nn>r.fn||C(r)}(r.on,r.mn,t),r.Jr=c(r.Jr,d(t)),r.En=F(r.on,0)}else n=function(r,n,f){return r.dr[((n&r.wr)<<r.hr)+((255&f)>>>8-r.hr)]}(r.Ur,h(r.Jr),r.En),r.En=7>r.state?function(r,n){var f=1;do{f=f<<1|$r(n,r.wn,f)}while(256>f);return f<<24>>24}(n,r.un):function(r,n,f){var t,o,u=1;do{if(o=f>>7&1,f<<=1,t=$r(n,r.wn,(1+o<<8)+u),u=u<<1|t,o!=t){for(;256>u;)u=u<<1|$r(n,r.wn,u);break}}while(256>u);return u<<24>>24}(n,r.un,F(r.on,r.mn)),function(r,n){r.rn[r.fn]=n,r.fn+=1,r.nn>r.fn||C(r)}(r.on,r.En),r.state=J(r.state),r.Jr=c(r.Jr,u);return 0}(r.Pr);if(-1==n)throw Error("corrupted input");r.Nn=f,r.$n=r.Pr.Jr,(n||v(r.Pr.yn,o)>=0&&v(r.Pr.Jr,r.Pr.yn)>=0)&&(C(r.Pr.on),G(r.Pr.on),r.Pr.un.Nr=null,r.Qr=0)}(r),r.Qr}function P(r){if(!r.Qr)throw Error("bad state");if(!r.Kr)throw Error("No decoding");return function(r){(function(r,n,f,t){var i,e,a,l,p,w,M,b,x,g,y,E,k,S;n[0]=o,f[0]=o,t[0]=1,r.er&&(r.p.pn=r.er,function(r){r.kn=0,r.fn=0,r.tn=0,r.Sn=0,T(r),r.jn=0,q(r,-1)}(r.p),r.ir=1,r.er=null);if(r.cr)return;if(r.cr=1,k=r.Jr,s(r.Jr,o)){if(!O(r.p))return void nr(r,h(r.Jr));ir(r),E=h(r.Jr)&r.D,Or(r.N,r.$,(r.$r<<4)+E,0),r.$r=J(r.$r),a=_(r.p,-r.Gr),wr(pr(r.H,h(r.Jr),r._r),r.N,a),r._r=a,r.Gr-=1,r.Jr=c(r.Jr,u)}if(!O(r.p))return void nr(r,h(r.Jr));for(;;){if(w=fr(r,h(r.Jr)),x=r.or,E=h(r.Jr)&r.D,e=(r.$r<<4)+E,1==w&&-1==x)Or(r.N,r.$,e,0),a=_(r.p,-r.Gr),S=pr(r.H,h(r.Jr),r._r),7>r.$r?wr(S,r.N,a):(b=_(r.p,-r.j[0]-1-r.Gr),Mr(S,r.N,b,a)),r._r=a,r.$r=J(r.$r);else{if(Or(r.N,r.$,e,1),4>x){if(Or(r.N,r._,r.$r,1),x?(Or(r.N,r.L,r.$r,1),1==x?Or(r.N,r.O,r.$r,0):(Or(r.N,r.O,r.$r,1),Or(r.N,r.T,r.$r,x-2))):(Or(r.N,r.L,r.$r,0),Or(r.N,r.q,e,1==w?0:1)),1==w?r.$r=7>r.$r?9:11:(lr(r.G,r.N,w-2,E),r.$r=7>r.$r?8:11),l=r.j[x],0!=x){for(var D=x;D>=1;--D)r.j[D]=r.j[D-1];r.j[0]=l}}else{Or(r.N,r._,r.$r,0),r.$r=7>r.$r?7:10,lr(r.F,r.N,w-2,E),y=cr(x-=4),M=H(w),kr(r.A[M],r.N,y),4>y||(g=x-(i=(2|1&y)<<(p=(y>>1)-1)),14>y?Ir(r.B,i-y-1,r.N,p,g):(Tr(r.N,g>>4,p-4),Dr(r.C,r.N,15&g),r._n+=1)),l=x;for(D=3;D>=1;--D)r.j[D]=r.j[D-1];r.j[0]=l,r.Ln+=1}r._r=_(r.p,w-1-r.Gr)}if(r.Gr-=w,r.Jr=c(r.Jr,d(w)),!r.Gr){if(128>r.Ln||rr(r),16>r._n||Z(r),n[0]=r.Jr,f[0]=qr(r.N),!O(r.p))return void nr(r,h(r.Jr));if(v(m(r.Jr,k),[4096,0])>=0)return r.cr=0,void(t[0]=0)}}})(r.Kr,r.Kr.W,r.Kr.X,r.Kr.Y),r.Nn=r.Kr.W[0],r.Kr.Y[0]&&(!function(r){er(r),r.N.Nr=null}(r.Kr),r.Qr=0)}(r),r.Qr}function Q(r,n){for(;n>r.On;r.On+=1)r.Tn[r.On]=gr({},3),r.qn[r.On]=gr({},3)}function R(r,n,f){if(!$r(n,r.zn,0))return yr(r.Tn[f],n);var t=8;return $r(n,r.zn,1)?t+=8+yr(r.An,n):t+=yr(r.qn[f],n),t}function U(r){return r.zn=e(2),r.Tn=e(16),r.qn=e(16),r.An=gr({},8),r.On=0,r}function V(r){_r(r.zn);for(var n=0;r.On>n;++n)_r(r.Tn[n].Ar),_r(r.qn[n].Ar);_r(r.An.Ar)}function W(r){return r.wn=e(768),r}var X=function(){var r,n,f,t=2,o=[0,1];for(f=2;22>f;++f){var u=f;for(u>>=1,n=1,n<<=u-=1,r=0;n>r;++r,++t)o[t]=f<<24>>24}return o}();function Y(r,n){var f,t,o,u;r.Cr=n,o=r.I[n].Bn,t=r.I[n].Cn;do{r.I[n].Fn&&(xr(r.I[o]),r.I[o].Bn=o-1,r.I[n].Gn&&(r.I[o-1].Fn=0,r.I[o-1].Bn=r.I[n].Hn,r.I[o-1].Cn=r.I[n].Jn)),f=t,t=r.I[u=o].Cn,o=r.I[u].Bn,r.I[u].Cn=f,r.I[u].Bn=n,n=u}while(n>0);return r.or=r.I[0].Cn,r.Fr=r.I[0].Bn,r.Fr}function Z(r){for(var n=0;16>n;++n)r.R[n]=jr(r.C,n);r._n=0}function rr(r){var n,f,t,o,u,i,e,c;for(o=4;128>o;++o)i=cr(o),r.Z[o]=Nr(r.B,(n=(2|1&i)<<(t=(i>>1)-1))-i-1,t,o-n);for(u=0;4>u;++u){for(f=r.A[u],e=u<<6,i=0;r.v>i;i+=1)r.K[e+i]=Sr(f,i);for(i=14;r.v>i;i+=1)r.K[e+i]+=(i>>1)-1-4<<6;for(c=128*u,o=0;4>o;++o)r.P[c+o]=r.K[e+o];for(;128>o;++o)r.P[c+o]=r.K[e+cr(o)]+r.Z[o]}r.Ln=0}function nr(r,n){er(r),function(r,n){Or(r.N,r.$,(r.$r<<4)+n,1),Or(r.N,r._,r.$r,0),r.$r=7>r.$r?7:10,lr(r.F,r.N,0,n);var f=H(2);kr(r.A[f],r.N,63),Tr(r.N,67108863,26),Dr(r.C,r.N,15)}(r,n&r.D);for(var f=0;5>f;++f)zr(r.N)}function fr(r,n){var f,t,o,u,i,e,c,a,v,l,s,d,h,p,w,M,m,b,x,g,y,E,k,S,D,j,I,N,$,T,q,z,A,B,C,F,G,H,K,P,Q,R,U;if(r.Cr!=r.Fr)return h=r.I[r.Fr].Bn-r.Fr,r.or=r.I[r.Fr].Cn,r.Fr=r.I[r.Fr].Bn,h;if(r.Fr=r.Cr=0,r.Br?(d=r.rr,r.Br=0):d=ir(r),j=r.nr,2>(S=O(r.p)+1))return r.or=-1,1;for(S>273&&(S=273),K=0,v=0;4>v;++v)r.U[v]=r.j[v],r.V[v]=L(r.p,-1,r.U[v],273),r.V[v]>r.V[K]&&(K=v);if(r.V[K]>=r.s)return r.or=K,ur(r,(h=r.V[K])-1),h;if(d>=r.s)return r.or=r.J[j-1]+4,ur(r,d-1),d;if(c=_(r.p,-1),m=_(r.p,-r.j[0]-1-1),2>d&&c!=m&&2>r.V[K])return r.or=-1,1;if(r.I[0].Kn=r.$r,r.I[1].Pn=Lr[r.$[(r.$r<<4)+(A=n&r.D)]>>>2]+br(pr(r.H,n,r._r),r.$r>=7,m,c),xr(r.I[1]),H=(b=Lr[2048-r.$[(r.$r<<4)+A]>>>2])+Lr[2048-r._[r.$r]>>>2],m==c&&(P=H+function(r,n,f){return Lr[r.L[n]>>>2]+Lr[r.q[(n<<4)+f]>>>2]}(r,r.$r,A),r.I[1].Pn>P&&(r.I[1].Pn=P,function(r){r.Cn=0,r.Fn=0}(r.I[1]))),2>(s=r.V[K]>d?r.V[K]:d))return r.or=r.I[1].Cn,1;r.I[1].Bn=0,r.I[0].Qn=r.U[0],r.I[0].Rn=r.U[1],r.I[0].Un=r.U[2],r.I[0].Vn=r.U[3],l=s;do{r.I[l].Pn=268435455,l-=1}while(l>=2);for(v=0;4>v;++v)if((G=r.V[v])>=2){C=H+or(r,v,r.$r,A);do{u=C+dr(r.G,G-2,A),(T=r.I[G]).Pn>u&&(T.Pn=u,T.Bn=0,T.Cn=v,T.Fn=0)}while((G-=1)>=2)}if(k=b+Lr[r._[r.$r]>>>2],d>=(l=2>r.V[0]?2:r.V[0]+1)){for(I=0;l>r.J[I];)I+=2;for(;u=k+tr(r,a=r.J[I+1],l,A),(T=r.I[l]).Pn>u&&(T.Pn=u,T.Bn=0,T.Cn=a+4,T.Fn=0),l!=r.J[I]||(I+=2)!=j;l+=1);}for(f=0;;){if(++f==s)return Y(r,f);if(x=ir(r),j=r.nr,x>=r.s)return r.rr=x,r.Br=1,Y(r,f);if(n+=1,z=r.I[f].Bn,r.I[f].Fn?(z-=1,r.I[f].Gn?(R=r.I[r.I[f].Hn].Kn,R=4>r.I[f].Jn?7>R?8:11:7>R?7:10):R=r.I[z].Kn,R=J(R)):R=r.I[z].Kn,z==f-1?R=r.I[f].Cn?J(R):7>R?9:11:(r.I[f].Fn&&r.I[f].Gn?(z=r.I[f].Hn,q=r.I[f].Jn,R=7>R?8:11):R=4>(q=r.I[f].Cn)?7>R?8:11:7>R?7:10,$=r.I[z],4>q?q?1==q?(r.U[0]=$.Rn,r.U[1]=$.Qn,r.U[2]=$.Un,r.U[3]=$.Vn):2==q?(r.U[0]=$.Un,r.U[1]=$.Qn,r.U[2]=$.Rn,r.U[3]=$.Vn):(r.U[0]=$.Vn,r.U[1]=$.Qn,r.U[2]=$.Rn,r.U[3]=$.Un):(r.U[0]=$.Qn,r.U[1]=$.Rn,r.U[2]=$.Un,r.U[3]=$.Vn):(r.U[0]=q-4,r.U[1]=$.Qn,r.U[2]=$.Rn,r.U[3]=$.Un)),r.I[f].Kn=R,r.I[f].Qn=r.U[0],r.I[f].Rn=r.U[1],r.I[f].Un=r.U[2],r.I[f].Vn=r.U[3],e=r.I[f].Pn,c=_(r.p,-1),m=_(r.p,-r.U[0]-1-1),t=e+Lr[r.$[(R<<4)+(A=n&r.D)]>>>2]+br(pr(r.H,n,_(r.p,-2)),R>=7,m,c),g=0,(y=r.I[f+1]).Pn>t&&(y.Pn=t,y.Bn=f,y.Cn=-1,y.Fn=0,g=1),H=(b=e+Lr[2048-r.$[(R<<4)+A]>>>2])+Lr[2048-r._[R]>>>2],m!=c||f>y.Bn&&!y.Cn||(P=H+(Lr[r.L[R]>>>2]+Lr[r.q[(R<<4)+A]>>>2]))>y.Pn||(y.Pn=P,y.Bn=f,y.Cn=0,y.Fn=0,g=1),(S=D=(D=O(r.p)+1)>4095-f?4095-f:D)>=2){if(S>r.s&&(S=r.s),!g&&m!=c&&(w=L(r.p,0,r.U[0],Math.min(D-1,r.s)))>=2){for(U=J(R),E=t+Lr[2048-r.$[(U<<4)+(B=n+1&r.D)]>>>2]+Lr[2048-r._[U]>>>2],N=f+1+w;N>s;)r.I[s+=1].Pn=268435455;u=E+(dr(r.G,w-2,B)+or(r,0,U,B)),(T=r.I[N]).Pn>u&&(T.Pn=u,T.Bn=f+1,T.Cn=0,T.Fn=1,T.Gn=0)}for(Q=2,F=0;4>F;++F)if((p=L(r.p,-1,r.U[F],S))>=2){M=p;do{for(;f+p>s;)r.I[s+=1].Pn=268435455;u=H+(dr(r.G,p-2,A)+or(r,F,R,A)),(T=r.I[f+p]).Pn>u&&(T.Pn=u,T.Bn=f,T.Cn=F,T.Fn=0)}while((p-=1)>=2);if(p=M,F||(Q=p+1),D>p&&(w=L(r.p,p,r.U[F],Math.min(D-1-p,r.s)))>=2){for(U=7>R?8:11,B=n+p&r.D,o=H+(dr(r.G,p-2,A)+or(r,F,R,A))+Lr[r.$[(U<<4)+B]>>>2]+br(pr(r.H,n+p,_(r.p,p-1-1)),1,_(r.p,p-1-(r.U[F]+1)),_(r.p,p-1)),U=J(U),E=o+Lr[2048-r.$[(U<<4)+(B=n+p+1&r.D)]>>>2]+Lr[2048-r._[U]>>>2],N=p+1+w;f+N>s;)r.I[s+=1].Pn=268435455;u=E+(dr(r.G,w-2,B)+or(r,0,U,B)),(T=r.I[f+N]).Pn>u&&(T.Pn=u,T.Bn=f+p+1,T.Cn=0,T.Fn=1,T.Gn=1,T.Hn=f,T.Jn=F)}}if(x>S){for(x=S,j=0;x>r.J[j];j+=2);r.J[j]=x,j+=2}if(x>=Q){for(k=b+Lr[r._[R]>>>2];f+x>s;)r.I[s+=1].Pn=268435455;for(I=0;Q>r.J[I];)I+=2;for(p=Q;;p+=1)if(u=k+tr(r,i=r.J[I+1],p,A),(T=r.I[f+p]).Pn>u&&(T.Pn=u,T.Bn=f,T.Cn=i+4,T.Fn=0),p==r.J[I]){if(D>p&&(w=L(r.p,p,i,Math.min(D-1-p,r.s)))>=2){for(o=u+Lr[r.$[((U=7>R?7:10)<<4)+(B=n+p&r.D)]>>>2]+br(pr(r.H,n+p,_(r.p,p-1-1)),1,_(r.p,p-(i+1)-1),_(r.p,p-1)),U=J(U),E=o+Lr[2048-r.$[(U<<4)+(B=n+p+1&r.D)]>>>2]+Lr[2048-r._[U]>>>2],N=p+1+w;f+N>s;)r.I[s+=1].Pn=268435455;u=E+(dr(r.G,w-2,B)+or(r,0,U,B)),(T=r.I[f+N]).Pn>u&&(T.Pn=u,T.Bn=f+p+1,T.Cn=0,T.Fn=1,T.Gn=1,T.Hn=f,T.Jn=i+4)}if((I+=2)==j)break}}}}}function tr(r,n,f,t){var o=H(f);return(128>n?r.P[128*o+n]:r.K[(o<<6)+function(r){if(131072>r)return X[r>>6]+12;if(134217728>r)return X[r>>16]+32;return X[r>>26]+52}(n)]+r.R[15&n])+dr(r.F,f-2,t)}function or(r,n,f,t){var o;return n?(o=Lr[2048-r.L[f]>>>2],1==n?o+=Lr[r.O[f]>>>2]:(o+=Lr[2048-r.O[f]>>>2],o+=Ar(r.T[f],n-2))):(o=Lr[r.L[f]>>>2],o+=Lr[2048-r.q[(f<<4)+t]>>>2]),o}function ur(r,n){n>0&&(!function(r,n){var f,t,o,u,i,e,c,a,v,l,s,d,h,p,w;do{if(r.tn<r.fn+r.Er){if(r.lr>(l=r.tn-r.fn)){A(r);continue}}else l=r.Er;for(s=r.fn>r.kr?r.fn-r.kr:0,t=r.kn+r.fn,r.ar?(r.Ir[1023&(w=z[255&r.xr[t]]^255&r.xr[t+1])]=r.fn,r.Ir[1024+(65535&(w^=(255&r.xr[t+2])<<8))]=r.fn,e=(w^z[255&r.xr[t+3]]<<5)&r.Dr):e=255&r.xr[t]^(255&r.xr[t+1])<<8,o=r.Ir[r.sr+e],r.Ir[r.sr+e]=r.fn,h=1+(r.jn<<1),p=r.jn<<1,a=v=r.vr,f=r.Mr;;){if(s>=o||0==f){f-=1,r.Sr[h]=r.Sr[p]=0;break}if(u=((i=r.fn-o)>r.jn?r.jn-i+r.kr:r.jn-i)<<1,r.xr[(d=r.kn+o)+(c=v>a?a:v)]==r.xr[t+c]){for(;(c+=1)!=l&&r.xr[d+c]==r.xr[t+c];);if(c==l){r.Sr[p]=r.Sr[u],r.Sr[h]=r.Sr[u+1];break}}(255&r.xr[t+c])>(255&r.xr[d+c])?(r.Sr[p]=o,o=r.Sr[p=u+1],v=c):(r.Sr[h]=o,o=r.Sr[h=u],a=c)}A(r)}while(0!=(n-=1))}(r.p,n),r.Gr+=n)}function ir(r){var n=0;return r.nr=function(r,n){var f,t,o,u,i,e,c,a,v,l,s,d,h,p,w,M,m,b,x,g,y;if(r.tn<r.fn+r.Er){if(r.lr>(p=r.tn-r.fn))return A(r),0}else p=r.Er;for(m=0,w=r.fn>r.kr?r.fn-r.kr:0,t=r.kn+r.fn,M=1,a=0,v=0,r.ar?(a=1023&(y=z[255&r.xr[t]]^255&r.xr[t+1]),v=65535&(y^=(255&r.xr[t+2])<<8),l=(y^z[255&r.xr[t+3]]<<5)&r.Dr):l=255&r.xr[t]^(255&r.xr[t+1])<<8,o=r.Ir[r.sr+l]||0,r.ar&&(u=r.Ir[a]||0,i=r.Ir[1024+v]||0,r.Ir[a]=r.fn,r.Ir[1024+v]=r.fn,u>w&&r.xr[r.kn+u]==r.xr[t]&&(n[m++]=M=2,n[m++]=r.fn-u-1),i>w&&r.xr[r.kn+i]==r.xr[t]&&(i==u&&(m-=2),n[m++]=M=3,n[m++]=r.fn-i-1,u=i),0!=m&&u==o&&(m-=2,M=1)),r.Ir[r.sr+l]=r.fn,x=1+(r.jn<<1),g=r.jn<<1,d=h=r.vr,0!=r.vr&&o>w&&r.xr[r.kn+o+r.vr]!=r.xr[t+r.vr]&&(n[m++]=M=r.vr,n[m++]=r.fn-o-1),f=r.Mr;;){if(w>=o||0==f){f-=1,r.Sr[x]=r.Sr[g]=0;break}if(e=((c=r.fn-o)>r.jn?r.jn-c+r.kr:r.jn-c)<<1,r.xr[(b=r.kn+o)+(s=h>d?d:h)]==r.xr[t+s]){for(;(s+=1)!=p&&r.xr[b+s]==r.xr[t+s];);if(s>M&&(n[m++]=M=s,n[m++]=c-1,s==p)){r.Sr[g]=r.Sr[e],r.Sr[x]=r.Sr[e+1];break}}(255&r.xr[t+s])>(255&r.xr[b+s])?(r.Sr[g]=o,o=r.Sr[g=e+1],h=s):(r.Sr[x]=o,o=r.Sr[x=e],d=s)}return A(r),m}(r.p,r.J),r.nr>0&&(n=r.J[r.nr-2])==r.s&&(n+=L(r.p,n-1,r.J[r.nr-1],273-n)),r.Gr+=1,n}function er(r){r.p&&r.ir&&(r.p.pn=null,r.ir=0)}function cr(r){return 2048>r?X[r]:2097152>r?X[r>>10]+20:X[r>>20]+40}function ar(r,n){_r(r.Wn);for(var f=0;n>f;++f)_r(r.Xn[f].Ar),_r(r.Yn[f].Ar);_r(r.Zn.Ar)}function vr(r,n,f,t,o){var u,i,e,c,a;for(u=Lr[r.Wn[0]>>>2],e=(i=Lr[2048-r.Wn[0]>>>2])+Lr[r.Wn[1]>>>2],c=i+Lr[2048-r.Wn[1]>>>2],a=0,a=0;8>a;++a){if(a>=f)return;t[o+a]=u+Sr(r.Xn[n],a)}for(;16>a;++a){if(a>=f)return;t[o+a]=e+Sr(r.Yn[n],a-8)}for(;f>a;++a)t[o+a]=c+Sr(r.Zn,a-8-8)}function lr(r,n,f,t){!function(r,n,f,t){8>f?(Or(n,r.Wn,0,0),kr(r.Xn[t],n,f)):(f-=8,Or(n,r.Wn,0,1),8>f?(Or(n,r.Wn,1,0),kr(r.Yn[t],n,f)):(Or(n,r.Wn,1,1),kr(r.Zn,n,f-8)))}(r,n,f,t),0==(r.rf[t]-=1)&&(vr(r,t,r.Hr,r.nf,272*t),r.rf[t]=r.Hr)}function sr(r){return function(r){r.Wn=e(2),r.Xn=e(16),r.Yn=e(16),r.Zn=Er({},8);for(var n=0;16>n;++n)r.Xn[n]=Er({},3),r.Yn[n]=Er({},3)}(r),r.nf=[],r.rf=[],r}function dr(r,n,f){return r.nf[272*f+n]}function hr(r,n){for(var f=0;n>f;++f)vr(r,f,r.Hr,r.nf,272*f),r.rf[f]=r.Hr}function pr(r,n,f){return r.dr[((n&r.wr)<<r.hr)+((255&f)>>>8-r.hr)]}function wr(r,n,f){var t,o,u=1;for(o=7;o>=0;--o)Or(n,r.zr,u,t=f>>o&1),u=u<<1|t}function Mr(r,n,f,t){var o,u,i,e,c=1,a=1;for(u=7;u>=0;--u)o=t>>u&1,e=a,c&&(e+=1+(i=f>>u&1)<<8,c=i==o),Or(n,r.zr,e,o),a=a<<1|o}function mr(r){return r.zr=e(768),r}function br(r,n,f,t){var o,u,i=1,e=7,c=0;if(n)for(;e>=0;--e)if(c+=Ar(r.zr[(1+(u=f>>e&1)<<8)+i],o=t>>e&1),i=i<<1|o,u!=o){--e;break}for(;e>=0;--e)c+=Ar(r.zr[i],o=t>>e&1),i=i<<1|o;return c}function xr(r){r.Cn=-1,r.Fn=0}function gr(r,n){return r.In=n,r.Ar=e(1<<n),r}function yr(r,n){var f,t=1;for(f=r.In;0!=f;f-=1)t=(t<<1)+$r(n,r.Ar,t);return t-(1<<r.In)}function Er(r,n){return r.In=n,r.Ar=e(1<<n),r}function kr(r,n,f){var t,o,u=1;for(o=r.In;0!=o;)Or(n,r.Ar,u,t=f>>>(o-=1)&1),u=u<<1|t}function Sr(r,n){var f,t,o=1,u=0;for(t=r.In;0!=t;)u+=Ar(r.Ar[o],f=n>>>(t-=1)&1),o=(o<<1)+f;return u}function Dr(r,n,f){var t,o,u=1;for(o=0;r.In>o;++o)Or(n,r.Ar,u,t=1&f),u=u<<1|t,f>>=1}function jr(r,n){var f,t,o=1,u=0;for(t=r.In;0!=t;t-=1)f=1&n,n>>>=1,u+=Ar(r.Ar[o],f),o=o<<1|f;return u}function Ir(r,n,f,t,o){var u,i,e=1;for(i=0;t>i;++i)Or(f,r,n+e,u=1&o),e=e<<1|u,o>>=1}function Nr(r,n,f,t){var o,u,i=1,e=0;for(u=f;0!=u;u-=1)o=1&t,t>>>=1,e+=Lr[(2047&(r[n+i]-o^-o))>>>2],i=i<<1|o;return e}function $r(r,n,f){var t,o=n[f];return(-2147483648^(t=(r.Range>>>11)*o))>(-2147483648^r.Mn)?(r.Range=t,n[f]=o+(2048-o>>>5)<<16>>16,-16777216&r.Range||(r.Mn=r.Mn<<8|x(r.Nr),r.Range<<=8),0):(r.Range-=t,r.Mn-=t,n[f]=o-(o>>>5)<<16>>16,-16777216&r.Range||(r.Mn=r.Mn<<8|x(r.Nr),r.Range<<=8),1)}function _r(r){for(var n=r.length-1;n>=0;--n)r[n]=1024}var Lr=function(){var r,n,f,t,o=[];for(n=8;n>=0;--n)for(t=1,r=1,r<<=9-n,f=t<<=9-n-1;r>f;++f)o[f]=(n<<6)+(r-f<<6>>>9-n-1);return o}();function Or(r,n,f,t){var o,u=n[f];o=(r.Range>>>11)*u,t?(r.Or=c(r.Or,a(d(o),[4294967295,0])),r.Range-=o,n[f]=u-(u>>>5)<<16>>16):(r.Range=o,n[f]=u+(2048-u>>>5)<<16>>16),-16777216&r.Range||(r.Range<<=8,zr(r))}function Tr(r,n,f){for(var t=f-1;t>=0;t-=1)r.Range>>>=1,1==(n>>>t&1)&&(r.Or=c(r.Or,d(r.Range))),-16777216&r.Range||(r.Range<<=8,zr(r))}function qr(r){return c(c(d(r.Tr),r.Lr),[4,0])}function zr(r){var n,f,t,o,u=h((t=32,o=M(f=r.Or,t&=63),0>f[1]&&(o=c(o,w([2,0],63-t))),o));if(0!=u||0>v(r.Or,[4278190080,0])){r.Lr=c(r.Lr,d(r.Tr)),n=r.qr;do{k(r.Nr,n+u),n=255}while(0!=(r.Tr-=1));r.qr=h(r.Or)>>>24}r.Tr+=1,r.Or=w(a(r.Or,[16777215,0]),8)}function Ar(r,n){return Lr[(2047&(r-n^-n))>>>2]}function Br(r){for(var n,f,t,o=0,u=0,i=r.length,e=[],c=[];i>o;++o,++u){if(128&(n=255&r[o]))if(192==(224&n)){if(o+1>=i)return r;if(128!=(192&(f=255&r[++o])))return r;c[u]=(31&n)<<6|63&f}else{if(224!=(240&n))return r;if(o+2>=i)return r;if(128!=(192&(f=255&r[++o])))return r;if(128!=(192&(t=255&r[++o])))return r;c[u]=(15&n)<<12|(63&f)<<6|63&t}else{if(!n)return r;c[u]=n}16383==u&&(e.push(String.fromCharCode.apply(String,c)),u=-1)}return u>0&&(c.length=u,e.push(String.fromCharCode.apply(String,c))),e.join("")}function Cr(r){var n,f,t,o=[],u=0,i=r.length;if("object"==typeof r)return r;for(function(r,n,f,t,o){var u;for(u=n;f>u;++u)t[o++]=r.charCodeAt(u)}(r,0,i,o,0),t=0;i>t;++t)1>(n=o[t])||n>127?u+=n&&(128>n||n>2047)?3:2:++u;for(f=[],u=0,t=0;i>t;++t)1>(n=o[t])||n>127?n&&(128>n||n>2047)?(f[u++]=(224|n>>12&15)<<24>>24,f[u++]=(128|n>>6&63)<<24>>24,f[u++]=(128|63&n)<<24>>24):(f[u++]=(192|n>>6&31)<<24>>24,f[u++]=(128|63&n)<<24>>24):f[u++]=n<<24>>24;return f}function Fr(r){return r[1]+r[0]}function compress(r,f,t,o){var u,e,c={},a=void 0===t&&void 0===o;if("function"!=typeof t&&(e=t,t=o=0),o=o||function(r){if(void 0!==e)return i(r,e)},t=t||function(r,n){if(void 0!==e)return postMessage({action:1,cbn:e,result:r,error:n})},a){for(c.c=I({},Cr(r),Hr(f));P(c.c.ur););return E(c.c.Rr)}try{c.c=I({},Cr(r),Hr(f)),o(0)}catch(r){return t(null,r)}n((function r(){try{for(var f,i=(new Date).getTime();P(c.c.ur);)if(u=Fr(c.c.ur.Nn)/Fr(c.c.u),(new Date).getTime()-i>200)return o(u),n(r,0),0;o(1),f=E(c.c.Rr),n(t.bind(null,f),0)}catch(r){t(null,r)}}),0)}function decompress(r,f,t){var o,u,e,c,a={},v=void 0===f&&void 0===t;if("function"!=typeof f&&(u=f,f=t=0),t=t||function(r){if(void 0!==u)return i(e?r:-1,u)},f=f||function(r,n){if(void 0!==u)return postMessage({action:2,cbn:u,result:r,error:n})},v){for(a.d=$({},r);K(a.d.ur););return Br(E(a.d.Rr))}try{a.d=$({},r),c=Fr(a.d.u),e=c>-1,t(0)}catch(r){return f(null,r)}n((function r(){try{for(var u,i=0,v=(new Date).getTime();K(a.d.ur);)if(++i%1e3==0&&(new Date).getTime()-v>200)return e&&(o=Fr(a.d.ur.Pr.Jr)/c,t(o)),n(r,0),0;t(1),u=Br(E(a.d.Rr)),n(f.bind(null,u),0)}catch(r){f(null,r)}}),0)}var Gr,Hr=(Gr=[{l:16,f:64,m:0},{l:20,f:64,m:0},{l:19,f:64,m:1},{l:20,f:64,m:1},{l:21,f:128,m:1},{l:22,f:128,m:1},{l:23,f:128,m:1},{l:24,f:255,m:1},{l:25,f:255,m:1}],function(r){return Gr[r-1]||Gr[6]}),LZMA=function(){};LZMA.compress=compress,LZMA.decompress=decompress,LZMA.prototype.compress=compress,LZMA.prototype.decompress=decompress,LZMA.compress=compress,LZMA.decompress=decompress,LZMA.prototype.compress=compress,LZMA.prototype.decompress=decompress,"undefined"!=typeof self&&"importScripts"in self&&addEventListener("message",(function(r){2==r.data.action?LZMA.decompress(r.data.data,r.data.cbn):1==r.data.action&&LZMA.compress(r.data.data,r.data.mode,r.data.cbn)})),r.LZMA=LZMA,r.compress=compress,r.decompress=decompress,Object.defineProperty(r,"ff",{value:!0})}));var LZMA=this.LZMA.LZMA,LZMA_WORKER=LZMA;
+"undefined"==typeof Worker||"undefined"!=typeof location&&"file:"===location.protocol?"undefined"!=typeof global&&"undefined"!=typeof require?this.LZMA=function(n){return require(n||"./lzma_worker.js").LZMA}:"undefined"!=typeof window&&window.document?function(){var n,e=this;function o(r){var t;return function(n){var r=document.createElement("script");r.type="text/javascript",r.src=n,r.onload=function(){e.LZMA=o},document.getElementsByTagName("head")[0].appendChild(r)}(r),t={compress:function(e,o,r,i){n.LZMA_WORKER?n.LZMA_WORKER.compress(e,o,r,i):setTimeout((function(){t.compress(e,o,r,i)}),50)},decompress:function(e,o,r){n.LZMA_WORKER?n.LZMA_WORKER.decompress(e,o,r):setTimeout((function(){t.decompress(e,o,r)}),50)},worker:function(){return null}}}"undefined"!=typeof window?n=window:global&&(n=global),e.LZMA=o}():console.error("Can't load the worker. Sorry."):this.LZMA=function(n){var e={},o=new Worker(n||"./lzma_worker-min.js");return o.onmessage=function(n){3===n.data.action?e[n.data.cbn]&&"function"==typeof e[n.data.cbn].on_progress&&e[n.data.cbn].on_progress(n.data.result):e[n.data.cbn]&&"function"==typeof e[n.data.cbn].on_finish&&(e[n.data.cbn].on_finish(n.data.result,n.data.error),delete e[n.data.cbn])},o.onerror=function(n){var o=new Error(n.message+" ("+n.filename+":"+n.lineno+")");for(var r in e)e[r].on_finish(null,o);console.error("Uncaught error in lzma_worker",o)},function(){function n(n,r,t,i,a){var c;do{c=Math.floor(1e7*Math.random())}while(void 0!==e[c]);e[c]={on_finish:i,on_progress:a},o.postMessage({action:n,cbn:c,data:r,mode:t})}return{compress:function(e,o,r,t){n(1,e,o,r,t)},decompress:function(e,o,r){n(2,e,!1,o,r)},worker:function(){return o}}}()};
