--- src/lzma-min.mjs
@@ -1,1 +1,0 @@
-var r="function"==typeof setImmediate?setImmediate:setTimeout,n=[4294967295,-4294967296],f=[0,-0x8000000000000000],t=[0,0],o=[1,0];function u(r,n){postMessage({action:3,cbn:n,result:r})}function i(r){var n=[];return n[r-1]=void 0,n}function e(r,n){return v(r[0]+n[0],r[1]+n[1])}function c(r,n){return function(r,n){var f;f=n,0>n&&(f+=4294967296);return[f,4294967296*r]}(~~Math.max(Math.min(r[1]/4294967296,2147483647),-2147483648)&~~Math.max(Math.min(n[1]/4294967296,2147483647),-2147483648),h(r)&h(n))}function a(r,n){var f,t;return r[0]==n[0]&&r[1]==n[1]?0:(t=0>n[1],(f=0>r[1])&&!t?-1:!f&&t?1:0>p(r,n)[1]?-1:1)}function v(r,n){var f,t;for(n=(n%=0x10000000000000000)-(f=n%4294967296)+(t=4294967296*Math.floor((r%=0x10000000000000000)/4294967296)),r=r-t+f;0>r;)r+=4294967296,n-=4294967296;for(;r>4294967295;)r-=4294967296,n+=4294967296;for(n%=0x10000000000000000;n>0x7fffffff00000000;)n-=0x10000000000000000;for(;-0x8000000000000000>n;)n+=0x10000000000000000;return[r,n]}function l(r,n){return r[0]==n[0]&&r[1]==n[1]}function s(r){return 0>r?[r+4294967296,-4294967296]:[r,0]}function h(r){return 2147483648>r[0]?~~Math.max(Math.min(r[0],2147483647),-2147483648):~~Math.max(Math.min(r[0]-4294967296,2147483647),-2147483648)}function d(r){return r>30?d(30)*d(r-30):1<<r}function w(r,n){var o,u,i,e;if(n&=63,l(r,f))return n?t:r;if(0>r[1])throw Error("Neg");return e=d(n),u=r[1]*e%0x10000000000000000,0x8000000000000000>(u+=o=(i=r[0]*e)-i%4294967296)||(u-=0x10000000000000000),[i-=o,u]}function M(r,n){var f;return f=d(n&=63),v(Math.floor(r[0]/f),r[1]/f)}function p(r,n){return v(r[0]-n[0],r[1]-n[1])}function m(r,n){return r.t=n,r.o=0,r.count=n.length,r}function b(r){return r.count>r.o?255&r.t[r.o++]:-1}function g(r,n,f,t){return r.count>r.o?(S(r.t,r.o,n,f,t=Math.min(t,r.count-r.o)),r.o+=t,t):-1}function x(r){return r.t=i(32),r.count=0,r}function E(r){var n=r.t;return n.length=r.count,n}function y(r,n){r.t[r.count++]=n<<24>>24}function k(r,n,f,t){S(n,f,r.t,r.count,t),r.count+=t}function S(r,n,f,t,o){for(var u=0;o>u;++u)f[t+u]=r[n+u]}function D(r,f,o,u,e){var c,v;if(0>a(u,n))throw Error("invalid length "+u);for(r.u=u,function(r,n){!function(r,n){r.i=n;for(var f=0;n>1<<f;++f);r.v=2*f}(n,1<<r.l),n.s=r.f,function(r,n){var f=r.h;r.h=n,r.M&&f!=r.h&&(r.p=-1,r.M=null)}(n,r.m),n.g=0,n.k=3,n.S=2,n.D=3}(e,c=function(r){var n;for(r.I=i(4),r.N=[],r.$={},r.j=i(192),r.L=i(12),r.T=i(12),r.q=i(12),r.A=i(12),r.B=i(192),r.C=[],r.F=i(114),r.G=Er({},4),r.H=lr({}),r.J=lr({}),r.K={},r.O=[],r.P=[],r.R=[],r.U=i(16),r.V=i(4),r.W=i(4),r.X=[t],r.Y=[t],r.Z=[0],r.properties=i(5),r._=i(128),r.rr=0,r.h=1,r.nr=0,r.tr=-1,r.or=0,n=0;4096>n;++n)r.N[n]={};for(n=0;4>n;++n)r.C[n]=Er({},6);return r}({})),function(r,n){r.properties[0]=9*(5*r.S+r.g)+r.k<<24>>24;for(var f=0;4>f;++f)r.properties[1+f]=r.i>>8*f<<24>>24;k(n,r.properties,0,5)}(c,o),v=0;64>v;v+=8)y(o,255&h(M(u,v)));r.ur=(c.ir=0,c.er=f,c.cr=0,function(r){var n,f;r.M||(f=4,r.h||(f=2),function(r,n){r.ar=n>2,r.ar?(r.vr=0,r.lr=4,r.sr=66560):(r.vr=2,r.lr=3,r.sr=0)}(n={},f),r.M=n);if(function(r,n,f){var t,o;if(null!=r.hr&&r.dr==f&&r.wr==n)return;for(r.wr=n,r.Mr=(1<<n)-1,r.dr=f,r.hr=i(o=1<<r.dr+r.wr),t=0;o>t;++t)r.hr[t]=pr({})}(r.K,r.g,r.k),r.i==r.p&&r.tr==r.s)return;(function(r,n,f,t,o){var u,e;1073741567>n&&(r.pr=16+(t>>1),function(r,n,f,t){var o;r.mr=n,r.br=f,o=n+f+t,(null==r.gr||r.xr!=o)&&(r.gr=null,r.xr=o,r.gr=i(r.xr));r.Er=r.xr-f}(r,n+f,t+o,256+~~((n+f+t+o)/2)),r.yr=t,r.kr!=(u=n+1)&&(r.Sr=i(2*(r.kr=u))),e=65536,r.ar&&(e=n-1,e|=e>>1,e|=e>>2,e|=e>>4,e|=e>>8,e>>=1,(e|=65535)>16777216&&(e>>=1),r.Dr=e,e+=1,e+=r.sr),e!=r.Ir&&(r.Nr=i(r.Ir=e)))})(r.M,r.i,4096,r.s,274),r.p=r.i,r.tr=r.s}(c),c.$.$r=o,function(r){(function(r){r.jr=0,r.Lr=0;for(var n=0;4>n;++n)r.I[n]=0})(r),function(r){r.Tr=t,r.qr=t,r.Range=-1,r.zr=1,r.Ar=0}(r.$),jr(r.j),jr(r.B),jr(r.L),jr(r.T),jr(r.q),jr(r.A),jr(r.F),function(r){var n,f=1<<r.dr+r.wr;for(n=0;f>n;++n)jr(r.hr[n].Br)}(r.K);for(var n=0;4>n;++n)jr(r.C[n].Cr);cr(r.H,1<<r.S),cr(r.J,1<<r.S),jr(r.G.Cr),r.Fr=0,r.Gr=0,r.Hr=0,r.Jr=0}(c),_(c),Z(c),c.H.Kr=c.s+1-2,hr(c.H,1<<c.S),c.J.Kr=c.s+1-2,hr(c.J,1<<c.S),c.Or=t,function(r,n){return r.Pr=n,r.Qr=null,r.Rr=1,r}({},c))}function I(r,n,f){return r.Ur=x({}),D(r,m({},n),r.Ur,s(n.length),f),r}function N(r,f,o){var u,e,c,a,v="",l=[];for(e=0;5>e;++e){if(-1==(c=b(f)))throw Error("truncated input");l[e]=c<<24>>24}if(!function(r,n){var f,t,o,u,e,c,a;if(5>n.length)return 0;for(o=(a=255&n[0])%9,u=(c=~~(a/9))%5,e=~~(c/5),f=0,t=0;4>t;++t)f+=(255&n[1+t])<<8*t;if(f>99999999||!function(r,n,f,t){if(n>8||f>4||t>4)return 0;!function(r,n,f){var t,o;if(null!=r.hr&&r.dr==f&&r.wr==n)return;for(r.wr=n,r.Mr=(1<<n)-1,r.dr=f,r.hr=i(o=1<<r.dr+r.wr),t=0;o>t;++t)r.hr[t]=W({})}(r.Vr,f,n);var o=1<<t;return Q(r.Wr,o),Q(r.Xr,o),r.Yr=o-1,1}(r,o,u,e))return 0;return function(r,n){if(0>n)return 0;r.Zr!=n&&(r.Zr=n,r._r=Math.max(r.Zr,1),function(r,n){null!=r.rn&&r.nn==n||(r.rn=i(n));r.nn=n,r.fn=0,r.tn=0}(r.on,Math.max(r._r,4096)));return 1}(r,f)}(u=function(r){r.on={},r.un={},r.in=i(192),r.en=i(12),r.cn=i(12),r.an=i(12),r.vn=i(12),r.ln=i(192),r.sn=i(4),r.hn=i(114),r.dn=gr({},4),r.Wr=U({}),r.Xr=U({}),r.Vr={};for(var n=0;4>n;++n)r.sn[n]=gr({},6);return r}({}),l))throw Error("corrupted input");for(e=0;64>e;e+=8){if(-1==(c=b(f)))throw Error("truncated input");1==(c=c.toString(16)).length&&(c="0"+c),v=c+""+v}r.u=/^0+$|^f+$/i.test(v)?n:(a=parseInt(v,16))>4294967295?n:s(a),r.ur=function(r,n,f,o){return r.un.$r=n,H(r.on),r.on.wn=f,function(r){r.on.tn=0,r.on.fn=0,jr(r.in),jr(r.ln),jr(r.en),jr(r.cn),jr(r.an),jr(r.vn),jr(r.hn),function(r){var n,f;for(f=1<<r.dr+r.wr,n=0;f>n;++n)jr(r.hr[n].Mn)}(r.Vr);for(var n=0;4>n;++n)jr(r.sn[n].Cr);V(r.Wr),V(r.Xr),jr(r.dn.Cr),function(r){r.pn=0,r.Range=-1;for(var n=0;5>n;++n)r.pn=r.pn<<8|b(r.$r)}(r.un)}(r),r.state=0,r.mn=0,r.bn=0,r.gn=0,r.xn=0,r.En=o,r.Or=t,r.yn=0,function(r,n){return r.Qr=n,r.Pr=null,r.Rr=1,r}({},r)}(u,f,o,r.u)}function $(r,n){return r.Ur=x({}),N(r,m({},n),r.Ur),r}function j(r,n){return r.gr[r.kn+r.fn+n]}function L(r,n,f,t){var o,u;for(r.Sn&&r.fn+n+t>r.tn&&(t=r.tn-(r.fn+n)),++f,u=r.kn+r.fn+n,o=0;t>o&&r.gr[u+o]==r.gr[u+o-f];++o);return o}function T(r){return r.tn-r.fn}function q(r){var n,f;if(!r.Sn)for(;;){if(!(f=-r.kn+r.xr-r.tn))return;if(-1==(n=g(r.wn,r.gr,r.kn+r.tn,f)))return r.Dn=r.tn,r.kn+r.Dn>r.Er&&(r.Dn=r.Er-r.kn),void(r.Sn=1);r.tn+=n,r.fn+r.br>r.tn||(r.Dn=r.tn-r.br)}}function z(r,n){r.kn+=n,r.Dn-=n,r.fn-=n,r.tn-=n}var A=function(){var r,n,f,t=[];for(r=0;256>r;++r){for(f=r,n=0;8>n;++n)0!=(1&f)?(f>>>=1,f^=-306674912):f>>>=1;t[r]=f}return t}();function B(r){var n;(r.In+=1)<r.kr||(r.In=0),function(r){r.fn+=1,r.fn>r.Dn&&(r.kn+r.fn>r.Er&&function(r){var n,f,t;for((t=r.kn+r.fn-r.mr)>0&&--t,f=r.kn+r.tn-t,n=0;f>n;++n)r.gr[n]=r.gr[t+n];r.kn-=t}(r),q(r))}(r),1073741823==r.fn&&(C(r.Sr,2*r.kr,n=r.fn-r.kr),C(r.Nr,r.Ir,n),z(r,n))}function C(r,n,f){var t,o;for(t=0;n>t;++t)(o=r[t]||0)>f?o-=f:o=0,r[t]=o}function F(r){var n=r.fn-r.tn;n&&(k(r.wn,r.rn,r.tn,n),r.nn>r.fn||(r.fn=0),r.tn=r.fn)}function G(r,n){var f=r.fn-n-1;return 0>f&&(f+=r.nn),r.rn[f]}function H(r){F(r),r.wn=null}function J(r){return 4>(r-=2)?r:3}function K(r){return 4>r?0:10>r?r-3:r-6}function O(r){if(!r.Rr)throw Error("bad state");if(r.Pr)throw Error("No encoding");return function(r){var f=function(r){var n,f,t,u,i,c;if(c=h(r.Or)&r.Yr,$r(r.un,r.in,(r.state<<4)+c)){if($r(r.un,r.en,r.state))t=0,$r(r.un,r.cn,r.state)?($r(r.un,r.an,r.state)?($r(r.un,r.vn,r.state)?(f=r.xn,r.xn=r.gn):f=r.gn,r.gn=r.bn):f=r.bn,r.bn=r.mn,r.mn=f):$r(r.un,r.ln,(r.state<<4)+c)||(r.state=7>r.state?9:11,t=1),t||(t=R(r.Xr,r.un,c)+2,r.state=7>r.state?8:11);else if(r.xn=r.gn,r.gn=r.bn,r.bn=r.mn,t=2+R(r.Wr,r.un,c),r.state=7>r.state?7:10,4>(i=xr(r.sn[J(t)],r.un)))r.mn=i;else if(r.mn=(2|1&i)<<(u=(i>>1)-1),14>i)r.mn+=function(r,n,f,t){var o,u,i=1,e=0;for(u=0;t>u;++u)o=$r(f,r,n+i),i<<=1,i+=o,e|=o<<u;return e}(r.hn,r.mn-i-1,r.un,u);else if(r.mn+=function(r,n){var f,t,o=0;for(f=n;0!=f;f-=1)r.Range>>>=1,r.pn-=r.Range&(t=r.pn-r.Range>>>31)-1,o=o<<1|1-t,-16777216&r.Range||(r.pn=r.pn<<8|b(r.$r),r.Range<<=8);return o}(r.un,u-4)<<4,r.mn+=function(r,n){var f,t,o=1,u=0;for(t=0;r.Nn>t;++t)f=$r(n,r.Cr,o),o<<=1,o+=f,u|=f<<t;return u}(r.dn,r.un),0>r.mn)return-1==r.mn?1:-1;if(a(s(r.mn),r.Or)>=0||r.mn>=r._r)return-1;!function(r,n,f){var t=r.fn-n-1;for(0>t&&(t+=r.nn);0!=f;f-=1)r.nn>t||(t=0),r.rn[r.fn]=r.rn[t],r.fn+=1,t+=1,r.nn>r.fn||F(r)}(r.on,r.mn,t),r.Or=e(r.Or,s(t)),r.yn=G(r.on,0)}else n=function(r,n,f){return r.hr[((n&r.Mr)<<r.dr)+((255&f)>>>8-r.dr)]}(r.Vr,h(r.Or),r.yn),r.yn=7>r.state?function(r,n){var f=1;do{f=f<<1|$r(n,r.Mn,f)}while(256>f);return f<<24>>24}(n,r.un):function(r,n,f){var t,o,u=1;do{if(o=f>>7&1,f<<=1,t=$r(n,r.Mn,(1+o<<8)+u),u=u<<1|t,o!=t){for(;256>u;)u=u<<1|$r(n,r.Mn,u);break}}while(256>u);return u<<24>>24}(n,r.un,G(r.on,r.mn)),function(r,n){r.rn[r.fn]=n,r.fn+=1,r.nn>r.fn||F(r)}(r.on,r.yn),r.state=K(r.state),r.Or=e(r.Or,o);return 0}(r.Qr);if(-1==f)throw Error("corrupted input");r.$n=n,r.jn=r.Qr.Or,(f||a(r.Qr.En,t)>=0&&a(r.Qr.Or,r.Qr.En)>=0)&&(F(r.Qr.on),H(r.Qr.on),r.Qr.un.$r=null,r.Rr=0)}(r),r.Rr}function P(r){if(!r.Rr)throw Error("bad state");if(!r.Pr)throw Error("No decoding");return function(r){(function(r,n,f,u){var i,c,v,d,w,M,m,b,g,x,E,y,k,S;n[0]=t,f[0]=t,u[0]=1,r.er&&(r.M.wn=r.er,function(r){r.kn=0,r.fn=0,r.tn=0,r.Sn=0,q(r),r.In=0,z(r,-1)}(r.M),r.ir=1,r.er=null);if(r.cr)return;if(r.cr=1,k=r.Or,l(r.Or,t)){if(!T(r.M))return void rr(r,h(r.Or));ur(r),y=h(r.Or)&r.D,Tr(r.$,r.j,(r.jr<<4)+y,0),r.jr=K(r.jr),v=j(r.M,-r.Jr),wr(dr(r.K,h(r.Or),r.Lr),r.$,v),r.Lr=v,r.Jr-=1,r.Or=e(r.Or,o)}if(!T(r.M))return void rr(r,h(r.Or));for(;;){if(M=nr(r,h(r.Or)),g=r.or,y=h(r.Or)&r.D,c=(r.jr<<4)+y,1==M&&-1==g)Tr(r.$,r.j,c,0),v=j(r.M,-r.Jr),S=dr(r.K,h(r.Or),r.Lr),7>r.jr?wr(S,r.$,v):(b=j(r.M,-r.I[0]-1-r.Jr),Mr(S,r.$,b,v)),r.Lr=v,r.jr=K(r.jr);else{if(Tr(r.$,r.j,c,1),4>g){if(Tr(r.$,r.L,r.jr,1),g?(Tr(r.$,r.T,r.jr,1),1==g?Tr(r.$,r.q,r.jr,0):(Tr(r.$,r.q,r.jr,1),Tr(r.$,r.A,r.jr,g-2))):(Tr(r.$,r.T,r.jr,0),Tr(r.$,r.B,c,1==M?0:1)),1==M?r.jr=7>r.jr?9:11:(vr(r.J,r.$,M-2,y),r.jr=7>r.jr?8:11),d=r.I[g],0!=g){for(var D=g;D>=1;--D)r.I[D]=r.I[D-1];r.I[0]=d}}else{Tr(r.$,r.L,r.jr,0),r.jr=7>r.jr?7:10,vr(r.H,r.$,M-2,y),E=er(g-=4),m=J(M),yr(r.C[m],r.$,E),4>E||(x=g-(i=(2|1&E)<<(w=(E>>1)-1)),14>E?Ir(r.F,i-E-1,r.$,w,x):(qr(r.$,x>>4,w-4),Sr(r.G,r.$,15&x),r.Ln+=1)),d=g;for(D=3;D>=1;--D)r.I[D]=r.I[D-1];r.I[0]=d,r.Tn+=1}r.Lr=j(r.M,M-1-r.Jr)}if(r.Jr-=M,r.Or=e(r.Or,s(M)),!r.Jr){if(128>r.Tn||_(r),16>r.Ln||Z(r),n[0]=r.Or,f[0]=zr(r.$),!T(r.M))return void rr(r,h(r.Or));if(a(p(r.Or,k),[4096,0])>=0)return r.cr=0,void(u[0]=0)}}})(r.Pr,r.Pr.X,r.Pr.Y,r.Pr.Z),r.$n=r.Pr.X[0],r.Pr.Z[0]&&(!function(r){ir(r),r.$.$r=null}(r.Pr),r.Rr=0)}(r),r.Rr}function Q(r,n){for(;n>r.qn;r.qn+=1)r.zn[r.qn]=gr({},3),r.An[r.qn]=gr({},3)}function R(r,n,f){if(!$r(n,r.Bn,0))return xr(r.zn[f],n);var t=8;return $r(n,r.Bn,1)?t+=8+xr(r.Cn,n):t+=xr(r.An[f],n),t}function U(r){return r.Bn=i(2),r.zn=i(16),r.An=i(16),r.Cn=gr({},8),r.qn=0,r}function V(r){jr(r.Bn);for(var n=0;r.qn>n;++n)jr(r.zn[n].Cr),jr(r.An[n].Cr);jr(r.Cn.Cr)}function W(r){return r.Mn=i(768),r}var X=function(){var r,n,f,t=2,o=[0,1];for(f=2;22>f;++f){var u=f;for(u>>=1,n=1,n<<=u-=1,r=0;n>r;++r,++t)o[t]=f<<24>>24}return o}();function Y(r,n){var f,t,o,u;r.Gr=n,o=r.N[n].Fn,t=r.N[n].Gn;do{r.N[n].Hn&&(br(r.N[o]),r.N[o].Fn=o-1,r.N[n].Jn&&(r.N[o-1].Hn=0,r.N[o-1].Fn=r.N[n].Kn,r.N[o-1].Gn=r.N[n].On)),f=t,t=r.N[u=o].Gn,o=r.N[u].Fn,r.N[u].Gn=f,r.N[u].Fn=n,n=u}while(n>0);return r.or=r.N[0].Gn,r.Hr=r.N[0].Fn,r.Hr}function Z(r){for(var n=0;16>n;++n)r.U[n]=Dr(r.G,n);r.Ln=0}function _(r){var n,f,t,o,u,i,e,c;for(o=4;128>o;++o)i=er(o),r._[o]=Nr(r.F,(n=(2|1&i)<<(t=(i>>1)-1))-i-1,t,o-n);for(u=0;4>u;++u){for(f=r.C[u],e=u<<6,i=0;r.v>i;i+=1)r.P[e+i]=kr(f,i);for(i=14;r.v>i;i+=1)r.P[e+i]+=(i>>1)-1-4<<6;for(c=128*u,o=0;4>o;++o)r.R[c+o]=r.P[e+o];for(;128>o;++o)r.R[c+o]=r.P[e+er(o)]+r._[o]}r.Tn=0}function rr(r,n){ir(r),function(r,n){Tr(r.$,r.j,(r.jr<<4)+n,1),Tr(r.$,r.L,r.jr,0),r.jr=7>r.jr?7:10,vr(r.H,r.$,0,n);var f=J(2);yr(r.C[f],r.$,63),qr(r.$,67108863,26),Sr(r.G,r.$,15)}(r,n&r.D);for(var f=0;5>f;++f)Ar(r.$)}function nr(r,n){var f,t,o,u,i,e,c,a,v,l,s,h,d,w,M,p,m,b,g,x,E,y,k,S,D,I,N,$,q,z,A,B,C,F,G,H,J,O,P,Q,R,U,V;if(r.Gr!=r.Hr)return d=r.N[r.Hr].Fn-r.Hr,r.or=r.N[r.Hr].Gn,r.Hr=r.N[r.Hr].Fn,d;if(r.Hr=r.Gr=0,r.Fr?(h=r.rr,r.Fr=0):h=ur(r),I=r.nr,2>(S=T(r.M)+1))return r.or=-1,1;for(S>273&&(S=273),P=0,v=0;4>v;++v)r.V[v]=r.I[v],r.W[v]=L(r.M,-1,r.V[v],273),r.W[v]>r.W[P]&&(P=v);if(r.W[P]>=r.s)return r.or=P,or(r,(d=r.W[P])-1),d;if(h>=r.s)return r.or=r.O[I-1]+4,or(r,h-1),h;if(c=j(r.M,-1),m=j(r.M,-r.I[0]-1-1),2>h&&c!=m&&2>r.W[P])return r.or=-1,1;if(r.N[0].Pn=r.jr,r.N[1].Qn=Lr[r.j[(r.jr<<4)+(C=n&r.D)]>>>2]+mr(dr(r.K,n,r.Lr),r.jr>=7,m,c),br(r.N[1]),O=(b=Lr[2048-r.j[(r.jr<<4)+C]>>>2])+Lr[2048-r.L[r.jr]>>>2],m==c&&(Q=O+function(r,n,f){return Lr[r.T[n]>>>2]+Lr[r.B[(n<<4)+f]>>>2]}(r,r.jr,C),r.N[1].Qn>Q&&(r.N[1].Qn=Q,function(r){r.Gn=0,r.Hn=0}(r.N[1]))),2>(s=r.W[P]>h?r.W[P]:h))return r.or=r.N[1].Gn,1;r.N[1].Fn=0,r.N[0].Rn=r.V[0],r.N[0].Un=r.V[1],r.N[0].Vn=r.V[2],r.N[0].Wn=r.V[3],l=s;do{r.N[l].Qn=268435455,l-=1}while(l>=2);for(v=0;4>v;++v)if((J=r.W[v])>=2){G=O+tr(r,v,r.jr,C);do{u=G+sr(r.J,J-2,C),(z=r.N[J]).Qn>u&&(z.Qn=u,z.Fn=0,z.Gn=v,z.Hn=0)}while((J-=1)>=2)}if(k=b+Lr[r.L[r.jr]>>>2],h>=(l=2>r.W[0]?2:r.W[0]+1)){for(N=0;l>r.O[N];)N+=2;for(;u=k+fr(r,a=r.O[N+1],l,C),(z=r.N[l]).Qn>u&&(z.Qn=u,z.Fn=0,z.Gn=a+4,z.Hn=0),l!=r.O[N]||(N+=2)!=I;l+=1);}for(f=0;;){if(++f==s)return Y(r,f);if(g=ur(r),I=r.nr,g>=r.s)return r.rr=g,r.Fr=1,Y(r,f);if(n+=1,B=r.N[f].Fn,r.N[f].Hn?(B-=1,r.N[f].Jn?(U=r.N[r.N[f].Kn].Pn,U=4>r.N[f].On?7>U?8:11:7>U?7:10):U=r.N[B].Pn,U=K(U)):U=r.N[B].Pn,B==f-1?U=r.N[f].Gn?K(U):7>U?9:11:(r.N[f].Hn&&r.N[f].Jn?(B=r.N[f].Kn,A=r.N[f].On,U=7>U?8:11):U=4>(A=r.N[f].Gn)?7>U?8:11:7>U?7:10,q=r.N[B],4>A?A?1==A?(r.V[0]=q.Un,r.V[1]=q.Rn,r.V[2]=q.Vn,r.V[3]=q.Wn):2==A?(r.V[0]=q.Vn,r.V[1]=q.Rn,r.V[2]=q.Un,r.V[3]=q.Wn):(r.V[0]=q.Wn,r.V[1]=q.Rn,r.V[2]=q.Un,r.V[3]=q.Vn):(r.V[0]=q.Rn,r.V[1]=q.Un,r.V[2]=q.Vn,r.V[3]=q.Wn):(r.V[0]=A-4,r.V[1]=q.Rn,r.V[2]=q.Un,r.V[3]=q.Vn)),r.N[f].Pn=U,r.N[f].Rn=r.V[0],r.N[f].Un=r.V[1],r.N[f].Vn=r.V[2],r.N[f].Wn=r.V[3],e=r.N[f].Qn,c=j(r.M,-1),m=j(r.M,-r.V[0]-1-1),t=e+Lr[r.j[(U<<4)+(C=n&r.D)]>>>2]+mr(dr(r.K,n,j(r.M,-2)),U>=7,m,c),x=0,(E=r.N[f+1]).Qn>t&&(E.Qn=t,E.Fn=f,E.Gn=-1,E.Hn=0,x=1),O=(b=e+Lr[2048-r.j[(U<<4)+C]>>>2])+Lr[2048-r.L[U]>>>2],m!=c||f>E.Fn&&!E.Gn||(Q=O+(Lr[r.T[U]>>>2]+Lr[r.B[(U<<4)+C]>>>2]))>E.Qn||(E.Qn=Q,E.Fn=f,E.Gn=0,E.Hn=0,x=1),(S=D=(D=T(r.M)+1)>4095-f?4095-f:D)>=2){if(S>r.s&&(S=r.s),!x&&m!=c&&(M=L(r.M,0,r.V[0],Math.min(D-1,r.s)))>=2){for(V=K(U),y=t+Lr[2048-r.j[(V<<4)+(F=n+1&r.D)]>>>2]+Lr[2048-r.L[V]>>>2],$=f+1+M;$>s;)r.N[s+=1].Qn=268435455;u=y+(sr(r.J,M-2,F)+tr(r,0,V,F)),(z=r.N[$]).Qn>u&&(z.Qn=u,z.Fn=f+1,z.Gn=0,z.Hn=1,z.Jn=0)}for(R=2,H=0;4>H;++H)if((w=L(r.M,-1,r.V[H],S))>=2){p=w;do{for(;f+w>s;)r.N[s+=1].Qn=268435455;u=O+(sr(r.J,w-2,C)+tr(r,H,U,C)),(z=r.N[f+w]).Qn>u&&(z.Qn=u,z.Fn=f,z.Gn=H,z.Hn=0)}while((w-=1)>=2);if(w=p,H||(R=w+1),D>w&&(M=L(r.M,w,r.V[H],Math.min(D-1-w,r.s)))>=2){for(V=7>U?8:11,F=n+w&r.D,o=O+(sr(r.J,w-2,C)+tr(r,H,U,C))+Lr[r.j[(V<<4)+F]>>>2]+mr(dr(r.K,n+w,j(r.M,w-1-1)),1,j(r.M,w-1-(r.V[H]+1)),j(r.M,w-1)),V=K(V),y=o+Lr[2048-r.j[(V<<4)+(F=n+w+1&r.D)]>>>2]+Lr[2048-r.L[V]>>>2],$=w+1+M;f+$>s;)r.N[s+=1].Qn=268435455;u=y+(sr(r.J,M-2,F)+tr(r,0,V,F)),(z=r.N[f+$]).Qn>u&&(z.Qn=u,z.Fn=f+w+1,z.Gn=0,z.Hn=1,z.Jn=1,z.Kn=f,z.On=H)}}if(g>S){for(g=S,I=0;g>r.O[I];I+=2);r.O[I]=g,I+=2}if(g>=R){for(k=b+Lr[r.L[U]>>>2];f+g>s;)r.N[s+=1].Qn=268435455;for(N=0;R>r.O[N];)N+=2;for(w=R;;w+=1)if(u=k+fr(r,i=r.O[N+1],w,C),(z=r.N[f+w]).Qn>u&&(z.Qn=u,z.Fn=f,z.Gn=i+4,z.Hn=0),w==r.O[N]){if(D>w&&(M=L(r.M,w,i,Math.min(D-1-w,r.s)))>=2){for(o=u+Lr[r.j[((V=7>U?7:10)<<4)+(F=n+w&r.D)]>>>2]+mr(dr(r.K,n+w,j(r.M,w-1-1)),1,j(r.M,w-(i+1)-1),j(r.M,w-1)),V=K(V),y=o+Lr[2048-r.j[(V<<4)+(F=n+w+1&r.D)]>>>2]+Lr[2048-r.L[V]>>>2],$=w+1+M;f+$>s;)r.N[s+=1].Qn=268435455;u=y+(sr(r.J,M-2,F)+tr(r,0,V,F)),(z=r.N[f+$]).Qn>u&&(z.Qn=u,z.Fn=f+w+1,z.Gn=0,z.Hn=1,z.Jn=1,z.Kn=f,z.On=i+4)}if((N+=2)==I)break}}}}}function fr(r,n,f,t){var o=J(f);return(128>n?r.R[128*o+n]:r.P[(o<<6)+function(r){if(131072>r)return X[r>>6]+12;if(134217728>r)return X[r>>16]+32;return X[r>>26]+52}(n)]+r.U[15&n])+sr(r.H,f-2,t)}function tr(r,n,f,t){var o;return n?(o=Lr[2048-r.T[f]>>>2],1==n?o+=Lr[r.q[f]>>>2]:(o+=Lr[2048-r.q[f]>>>2],o+=Br(r.A[f],n-2))):(o=Lr[r.T[f]>>>2],o+=Lr[2048-r.B[(f<<4)+t]>>>2]),o}function or(r,n){n>0&&(!function(r,n){var f,t,o,u,i,e,c,a,v,l,s,h,d,w,M;do{if(r.tn<r.fn+r.yr){if(r.lr>(l=r.tn-r.fn)){B(r);continue}}else l=r.yr;for(s=r.fn>r.kr?r.fn-r.kr:0,t=r.kn+r.fn,r.ar?(r.Nr[1023&(M=A[255&r.gr[t]]^255&r.gr[t+1])]=r.fn,r.Nr[1024+(65535&(M^=(255&r.gr[t+2])<<8))]=r.fn,e=(M^A[255&r.gr[t+3]]<<5)&r.Dr):e=255&r.gr[t]^(255&r.gr[t+1])<<8,o=r.Nr[r.sr+e],r.Nr[r.sr+e]=r.fn,d=1+(r.In<<1),w=r.In<<1,a=v=r.vr,f=r.pr;;){if(s>=o||0==f){f-=1,r.Sr[d]=r.Sr[w]=0;break}if(u=((i=r.fn-o)>r.In?r.In-i+r.kr:r.In-i)<<1,r.gr[(h=r.kn+o)+(c=v>a?a:v)]==r.gr[t+c]){for(;(c+=1)!=l&&r.gr[h+c]==r.gr[t+c];);if(c==l){r.Sr[w]=r.Sr[u],r.Sr[d]=r.Sr[u+1];break}}(255&r.gr[t+c])>(255&r.gr[h+c])?(r.Sr[w]=o,o=r.Sr[w=u+1],v=c):(r.Sr[d]=o,o=r.Sr[d=u],a=c)}B(r)}while(0!=(n-=1))}(r.M,n),r.Jr+=n)}function ur(r){var n=0;return r.nr=function(r,n){var f,t,o,u,i,e,c,a,v,l,s,h,d,w,M,p,m,b,g,x,E;if(r.tn<r.fn+r.yr){if(r.lr>(w=r.tn-r.fn))return B(r),0}else w=r.yr;for(m=0,M=r.fn>r.kr?r.fn-r.kr:0,t=r.kn+r.fn,p=1,a=0,v=0,r.ar?(a=1023&(E=A[255&r.gr[t]]^255&r.gr[t+1]),v=65535&(E^=(255&r.gr[t+2])<<8),l=(E^A[255&r.gr[t+3]]<<5)&r.Dr):l=255&r.gr[t]^(255&r.gr[t+1])<<8,o=r.Nr[r.sr+l]||0,r.ar&&(u=r.Nr[a]||0,i=r.Nr[1024+v]||0,r.Nr[a]=r.fn,r.Nr[1024+v]=r.fn,u>M&&r.gr[r.kn+u]==r.gr[t]&&(n[m++]=p=2,n[m++]=r.fn-u-1),i>M&&r.gr[r.kn+i]==r.gr[t]&&(i==u&&(m-=2),n[m++]=p=3,n[m++]=r.fn-i-1,u=i),0!=m&&u==o&&(m-=2,p=1)),r.Nr[r.sr+l]=r.fn,g=1+(r.In<<1),x=r.In<<1,h=d=r.vr,0!=r.vr&&o>M&&r.gr[r.kn+o+r.vr]!=r.gr[t+r.vr]&&(n[m++]=p=r.vr,n[m++]=r.fn-o-1),f=r.pr;;){if(M>=o||0==f){f-=1,r.Sr[g]=r.Sr[x]=0;break}if(e=((c=r.fn-o)>r.In?r.In-c+r.kr:r.In-c)<<1,r.gr[(b=r.kn+o)+(s=d>h?h:d)]==r.gr[t+s]){for(;(s+=1)!=w&&r.gr[b+s]==r.gr[t+s];);if(s>p&&(n[m++]=p=s,n[m++]=c-1,s==w)){r.Sr[x]=r.Sr[e],r.Sr[g]=r.Sr[e+1];break}}(255&r.gr[t+s])>(255&r.gr[b+s])?(r.Sr[x]=o,o=r.Sr[x=e+1],d=s):(r.Sr[g]=o,o=r.Sr[g=e],h=s)}return B(r),m}(r.M,r.O),r.nr>0&&(n=r.O[r.nr-2])==r.s&&(n+=L(r.M,n-1,r.O[r.nr-1],273-n)),r.Jr+=1,n}function ir(r){r.M&&r.ir&&(r.M.wn=null,r.ir=0)}function er(r){return 2048>r?X[r]:2097152>r?X[r>>10]+20:X[r>>20]+40}function cr(r,n){jr(r.Xn);for(var f=0;n>f;++f)jr(r.Yn[f].Cr),jr(r.Zn[f].Cr);jr(r._n.Cr)}function ar(r,n,f,t,o){var u,i,e,c,a;for(u=Lr[r.Xn[0]>>>2],e=(i=Lr[2048-r.Xn[0]>>>2])+Lr[r.Xn[1]>>>2],c=i+Lr[2048-r.Xn[1]>>>2],a=0,a=0;8>a;++a){if(a>=f)return;t[o+a]=u+kr(r.Yn[n],a)}for(;16>a;++a){if(a>=f)return;t[o+a]=e+kr(r.Zn[n],a-8)}for(;f>a;++a)t[o+a]=c+kr(r._n,a-8-8)}function vr(r,n,f,t){!function(r,n,f,t){8>f?(Tr(n,r.Xn,0,0),yr(r.Yn[t],n,f)):(f-=8,Tr(n,r.Xn,0,1),8>f?(Tr(n,r.Xn,1,0),yr(r.Zn[t],n,f)):(Tr(n,r.Xn,1,1),yr(r._n,n,f-8)))}(r,n,f,t),0==(r.rf[t]-=1)&&(ar(r,t,r.Kr,r.nf,272*t),r.rf[t]=r.Kr)}function lr(r){return function(r){r.Xn=i(2),r.Yn=i(16),r.Zn=i(16),r._n=Er({},8);for(var n=0;16>n;++n)r.Yn[n]=Er({},3),r.Zn[n]=Er({},3)}(r),r.nf=[],r.rf=[],r}function sr(r,n,f){return r.nf[272*f+n]}function hr(r,n){for(var f=0;n>f;++f)ar(r,f,r.Kr,r.nf,272*f),r.rf[f]=r.Kr}function dr(r,n,f){return r.hr[((n&r.Mr)<<r.dr)+((255&f)>>>8-r.dr)]}function wr(r,n,f){var t,o,u=1;for(o=7;o>=0;--o)Tr(n,r.Br,u,t=f>>o&1),u=u<<1|t}function Mr(r,n,f,t){var o,u,i,e,c=1,a=1;for(u=7;u>=0;--u)o=t>>u&1,e=a,c&&(e+=1+(i=f>>u&1)<<8,c=i==o),Tr(n,r.Br,e,o),a=a<<1|o}function pr(r){return r.Br=i(768),r}function mr(r,n,f,t){var o,u,i=1,e=7,c=0;if(n)for(;e>=0;--e)if(c+=Br(r.Br[(1+(u=f>>e&1)<<8)+i],o=t>>e&1),i=i<<1|o,u!=o){--e;break}for(;e>=0;--e)c+=Br(r.Br[i],o=t>>e&1),i=i<<1|o;return c}function br(r){r.Gn=-1,r.Hn=0}function gr(r,n){return r.Nn=n,r.Cr=i(1<<n),r}function xr(r,n){var f,t=1;for(f=r.Nn;0!=f;f-=1)t=(t<<1)+$r(n,r.Cr,t);return t-(1<<r.Nn)}function Er(r,n){return r.Nn=n,r.Cr=i(1<<n),r}function yr(r,n,f){var t,o,u=1;for(o=r.Nn;0!=o;)Tr(n,r.Cr,u,t=f>>>(o-=1)&1),u=u<<1|t}function kr(r,n){var f,t,o=1,u=0;for(t=r.Nn;0!=t;)u+=Br(r.Cr[o],f=n>>>(t-=1)&1),o=(o<<1)+f;return u}function Sr(r,n,f){var t,o,u=1;for(o=0;r.Nn>o;++o)Tr(n,r.Cr,u,t=1&f),u=u<<1|t,f>>=1}function Dr(r,n){var f,t,o=1,u=0;for(t=r.Nn;0!=t;t-=1)f=1&n,n>>>=1,u+=Br(r.Cr[o],f),o=o<<1|f;return u}function Ir(r,n,f,t,o){var u,i,e=1;for(i=0;t>i;++i)Tr(f,r,n+e,u=1&o),e=e<<1|u,o>>=1}function Nr(r,n,f,t){var o,u,i=1,e=0;for(u=f;0!=u;u-=1)o=1&t,t>>>=1,e+=Lr[(2047&(r[n+i]-o^-o))>>>2],i=i<<1|o;return e}function $r(r,n,f){var t,o=n[f];return(-2147483648^(t=(r.Range>>>11)*o))>(-2147483648^r.pn)?(r.Range=t,n[f]=o+(2048-o>>>5)<<16>>16,-16777216&r.Range||(r.pn=r.pn<<8|b(r.$r),r.Range<<=8),0):(r.Range-=t,r.pn-=t,n[f]=o-(o>>>5)<<16>>16,-16777216&r.Range||(r.pn=r.pn<<8|b(r.$r),r.Range<<=8),1)}function jr(r){for(var n=r.length-1;n>=0;--n)r[n]=1024}var Lr=function(){var r,n,f,t,o=[];for(n=8;n>=0;--n)for(t=1,r=1,r<<=9-n,f=t<<=9-n-1;r>f;++f)o[f]=(n<<6)+(r-f<<6>>>9-n-1);return o}();function Tr(r,n,f,t){var o,u=n[f];o=(r.Range>>>11)*u,t?(r.qr=e(r.qr,c(s(o),[4294967295,0])),r.Range-=o,n[f]=u-(u>>>5)<<16>>16):(r.Range=o,n[f]=u+(2048-u>>>5)<<16>>16),-16777216&r.Range||(r.Range<<=8,Ar(r))}function qr(r,n,f){for(var t=f-1;t>=0;t-=1)r.Range>>>=1,1==(n>>>t&1)&&(r.qr=e(r.qr,s(r.Range))),-16777216&r.Range||(r.Range<<=8,Ar(r))}function zr(r){return e(e(s(r.zr),r.Tr),[4,0])}function Ar(r){var n,f,t,o,u=h((t=32,o=M(f=r.qr,t&=63),0>f[1]&&(o=e(o,w([2,0],63-t))),o));if(0!=u||0>a(r.qr,[4278190080,0])){r.Tr=e(r.Tr,s(r.zr)),n=r.Ar;do{y(r.$r,n+u),n=255}while(0!=(r.zr-=1));r.Ar=h(r.qr)>>>24}r.zr+=1,r.qr=w(c(r.qr,[16777215,0]),8)}function Br(r,n){return Lr[(2047&(r-n^-n))>>>2]}function Cr(r){for(var n,f,t,o=0,u=0,i=r.length,e=[],c=[];i>o;++o,++u){if(128&(n=255&r[o]))if(192==(224&n)){if(o+1>=i)return r;if(128!=(192&(f=255&r[++o])))return r;c[u]=(31&n)<<6|63&f}else{if(224!=(240&n))return r;if(o+2>=i)return r;if(128!=(192&(f=255&r[++o])))return r;if(128!=(192&(t=255&r[++o])))return r;c[u]=(15&n)<<12|(63&f)<<6|63&t}else{if(!n)return r;c[u]=n}16383==u&&(e.push(String.fromCharCode.apply(String,c)),u=-1)}return u>0&&(c.length=u,e.push(String.fromCharCode.apply(String,c))),e.join("")}function Fr(r){var n,f,t,o=[],u=0,i=r.length;if("object"==typeof r)return r;for(function(r,n,f,t,o){var u;for(u=n;f>u;++u)t[o++]=r.charCodeAt(u)}(r,0,i,o,0),t=0;i>t;++t)1>(n=o[t])||n>127?u+=n&&(128>n||n>2047)?3:2:++u;for(f=[],u=0,t=0;i>t;++t)1>(n=o[t])||n>127?n&&(128>n||n>2047)?(f[u++]=(224|n>>12&15)<<24>>24,f[u++]=(128|n>>6&63)<<24>>24,f[u++]=(128|63&n)<<24>>24):(f[u++]=(192|n>>6&31)<<24>>24,f[u++]=(128|63&n)<<24>>24):f[u++]=n<<24>>24;return f}function Gr(r){return r[1]+r[0]}function compress(n,f,t,o){var i,e,c={},a=void 0===t&&void 0===o;if("function"!=typeof t&&(e=t,t=o=0),o=o||function(r){if(void 0!==e)return u(r,e)},t=t||function(r,n){if(void 0!==e)return postMessage({action:1,cbn:e,result:r,error:n})},a){for(c.c=I({},Fr(n),Jr(f));P(c.c.ur););return E(c.c.Ur)}try{c.c=I({},Fr(n),Jr(f)),o(0)}catch(r){return t(null,r)}r((function n(){try{for(var f,u=(new Date).getTime();P(c.c.ur);)if(i=Gr(c.c.ur.$n)/Gr(c.c.u),(new Date).getTime()-u>200)return o(i),r(n,0),0;o(1),f=E(c.c.Ur),r(t.bind(null,f),0)}catch(r){t(null,r)}}),0)}function decompress(n,f,t){var o,i,e,c,a={},v=void 0===f&&void 0===t;if("function"!=typeof f&&(i=f,f=t=0),t=t||function(r){if(void 0!==i)return u(e?r:-1,i)},f=f||function(r,n){if(void 0!==i)return postMessage({action:2,cbn:i,result:r,error:n})},v){for(a.d=$({},n);O(a.d.ur););return Cr(E(a.d.Ur))}try{a.d=$({},n),c=Gr(a.d.u),e=c>-1,t(0)}catch(r){return f(null,r)}r((function n(){try{for(var u,i=0,v=(new Date).getTime();O(a.d.ur);)if(++i%1e3==0&&(new Date).getTime()-v>200)return e&&(o=Gr(a.d.ur.Qr.Or)/c,t(o)),r(n,0),0;t(1),u=Cr(E(a.d.Ur)),r(f.bind(null,u),0)}catch(r){f(null,r)}}),0)}var Hr,Jr=(Hr=[{l:16,f:64,m:0},{l:20,f:64,m:0},{l:19,f:64,m:1},{l:20,f:64,m:1},{l:21,f:128,m:1},{l:22,f:128,m:1},{l:23,f:128,m:1},{l:24,f:255,m:1},{l:25,f:255,m:1}],function(r){return Hr[r-1]||Hr[6]}),LZMA=function(){};LZMA.compress=compress,LZMA.decompress=decompress,LZMA.prototype.compress=compress,LZMA.prototype.decompress=decompress,LZMA.compress=compress,LZMA.decompress=decompress,LZMA.prototype.compress=compress,LZMA.prototype.decompress=decompress,"undefined"!=typeof self&&"importScripts"in self&&addEventListener("message",(function(r){2==r.data.action?LZMA.decompress(r.data.data,r.data.cbn):1==r.data.action&&LZMA.compress(r.data.data,r.data.mode,r.data.cbn)}));export{LZMA,compress,decompress};
